refactor(graphql-client): перенести normalizeUri в utils
Jira issue: VEGA-911
